[feat] Refator files and Separate by Folder #89
