--- i-scheduler/i-scheduler/Enums.swift
@@ -1,13 +1,0 @@
-//
-//  Enums.swift
-//  i-scheduler
-//
-//  Created by 권은빈 on 2021/12/16.
-//
-
-import Foundation
-
-enum ToolBarAction {
-    case add
-    case edit
-}
